# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentacion API - Aulas",
        "description": "Listado de endpoints de producto - explicacion",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://apiaulas.test/"
        }
    ],
    "paths": {
        "/api/categoria_productos": {
            "get": {
                "tags": [
                    "Categoria"
                ],
                "summary": "Listado de las categorías de productos que ofrece la tienda.",
                "operationId": "3378f83017547235f6a59925d2fa480f",
                "responses": {
                    "200": {
                        "description": "Lista de categorías",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "rows": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id_categoria": {
                                                        "type": "number",
                                                        "example": "1"
                                                    },
                                                    "nombre_categoria": {
                                                        "type": "string",
                                                        "example": "Cuadernos"
                                                    },
                                                    "descripcion": {
                                                        "type": "string",
                                                        "example": "Descripción que puede ser cuadernos, lápices, etc."
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Categoria"
                ],
                "summary": "Registrar entrada",
                "operationId": "5df53f72e4ec2dbb6f9a454aff911ce8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre_categoria",
                                    "descripcion"
                                ],
                                "properties": {
                                    "nombre_categoria": {
                                        "type": "string",
                                        "example": "Categoria de tijeras (ejemplo)."
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "example": "Describiendo algún aspecto importante de la categoría tijeras."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Categoría registrada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Categoría registrada exitosamente"
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": 201
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Datos de entrada no válidos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Algo falló al intentar registrar la categoría"
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": 422
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/categoria_productos/{id_categoria}": {
            "get": {
                "tags": [
                    "Categoria"
                ],
                "summary": "Mostrar la información de un pedido en específico.",
                "operationId": "dc83fcd7db54f1e9179558973c795d75",
                "parameters": [
                    {
                        "name": "id_categoria",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Categoría retornado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id_categoria": {
                                            "type": "number",
                                            "example": 3
                                        },
                                        "nombre_categoria": {
                                            "type": "string",
                                            "example": "Nombre de categoría ejemplo lápices"
                                        },
                                        "descripcion": {
                                            "type": "string",
                                            "example": "Describiendo la categoría ejemplo lápices"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "NOT FOUND",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No query results for model [App\\\\Models\\\\Pedido] #id_pedido"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Categoria"
                ],
                "summary": "Actualizar categoría",
                "description": "Actualizar categoría existente",
                "operationId": "c3e8e1fbba5c33289874f00586066d9e",
                "parameters": [
                    {
                        "name": "id_categoria",
                        "in": "path",
                        "description": "ID de la categoría a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre_categoria": {
                                        "type": "string",
                                        "example": "Nuevo nombre de categoría"
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "example": "Nueva descripción de la categoría"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Categoría actualizada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Categoría actualizada exitosamente"
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": 200
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Categoría no encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "La categoría no fue encontrada"
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": 404
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Datos de entrada no válidos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Algo falló al actualizar la categoría"
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": 422
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Categoria"
                ],
                "summary": "Eliminar categoría",
                "description": "Eliminar una categoría existente",
                "operationId": "14bf61eb8d07997bf06feb0b26f4e533",
                "parameters": [
                    {
                        "name": "id_categoria",
                        "in": "path",
                        "description": "ID de la categoría a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Categoría eliminada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Categoría eliminada exitosamente"
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": 200
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Categoría no encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "La categoría no fue encontrada"
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": 404
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}